Moves card tests
